Refactor/#85: 게시글 전체(페이징) 조회 기능 에러로 인한 코드 수정

+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/board/entity/Board.java
@@ -62,7 +62,7 @@ public class Board extends BaseTimeEntity {
 	private boolean isDeleted;
 
 	@Column(nullable = false, name = "view_count")
-	private int viewCount;
+	private long viewCount;
 
 	public Board(User user, BoardType boardType, CategoryType category, String title, String content, int clover) {
 		this.user = user;
